% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FrantiClusters.R
\docType{data}
\name{FrantiClusters}
\alias{FrantiClusters}
\title{Clustering basic benchmarks from Fränti and Sieranoja}
\format{
Each sample is a list containing:

\describe{
\item{values}{The position (x,y) of each element.}
\item{k}{The suggested number of clusters.}
}
}
\source{
\url{http://cs.joensuu.fi/sipu/datasets/}
}
\usage{
FrantiClusters
}
\description{
Some of the custering basic benchmarks from

Fränti, P., & Sieranoja, S. (2018). K-means properties on six clustering benchmark datasets. Applied Intelligence, 48(12), 4743-4759.
}
\examples{

FrantiClusters$s1

}
\keyword{datasets}
