% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocalSearchFSFunctions.R
\name{ILSFS}
\alias{ILSFS}
\title{Iterated local search for the permutative flowshop}
\usage{
ILSFS(M, ni = 4, iter = 100, opt = "HC", ...)
}
\arguments{
\item{M}{A matrix object storing the time to process task j (column) in machine i (row).}

\item{ni}{Number of insertions performed when perturbating the solution.}

\item{iter}{number of iterations.}

\item{opt}{the local search algorithm to implement. Presently are supported \code{HC} (hill climbing with \link{HCFS}), \code{SA} (simulated annealing with \link{SAFS}) and \code{TS} (tabu search with \link{TSFS}).}

\item{...}{we can pass additional parameters for the \code{SAFS} heuristic if desired (\code{Tmax} and \code{mu}).}
}
\value{
sol The obtained solution.

obj The makespan of the obtained solution.
}
\description{
An iterated local search for the permutative flowshop. At each iteration, the present examined solution obtained at the previous iteration is perturbated performing \code{ni} insertions. This perturbated solution is improved with a local search heuristic using the swap operator. The starting solution is obtained with the Palmer heuristic.
}
\examples{
set.seed(2020)
instance <- matrix(sample(10:90, 100, replace=TRUE), 5, 20)
ILSFS(M=instance, opt="SA", Tmax=10000, mu=1000)
}
