% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocalSearchFSFunctions.R
\name{SAFS}
\alias{SAFS}
\title{Simulated annealing heuristic for permutative flowshop}
\usage{
SAFS(M, inisol, Tmax = 1000, mu = 1, op = "swap", eval = FALSE)
}
\arguments{
\item{M}{A matrix object storing the time to process task j (column) in machine i (row).}

\item{inisol}{The starting solution.}

\item{Tmax}{the maximum value of temperature, representing the number of iterations of the algorithm.}

\item{mu}{a parameter tuning the value of probability of selection a solution worse than the presently explored. Higher values of \code{mu} reduce the probability of selecting a worse solution.}

\item{op}{The neighbourhood operator. Presently are implemented \code{swap} (the default) and \code{insertion} neighbourhoods.}

\item{eval}{if set to \code{TRUE} returns information about the evolution of the algorithm.}
}
\value{
sol The obtained solution.

obj The makespan of the obtained solution.

evaltest (returned if \code{eval=TRUE}) the value of the tested solution in each iteration.

evalfit (returned if \code{eval=TRUE})  the value of the makespan of the examined solution in each iteration.

evalbest (returned if \code{eval=TRUE}) the value of the best solution obtained in each iteration.
}
\description{
An implementation of a simulated annealing heuristic for the flowshop.
}
\examples{
set.seed(2020)
instance <- matrix(sample(10:90, 100, replace=TRUE), 5, 20)
SAFS(M=instance, inisol=PalmerTrapezes(instance)$tra, Tmax=10000, mu=1000, op="insertion")
}
