% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WineQuality.R
\docType{data}
\name{WineQuality}
\alias{WineQuality}
\title{Wine quality data set}
\format{
A list with two data frames for red and white wine. As the samples are proposed for clustering, the quality variable is introduced and the rest of variables centered and scaled.
}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/wine+quality}
}
\usage{
WineQuality
}
\description{
Two datasets with chemical properties of red and white vinho verde wine samples, from the north of Portugal.
}
\examples{

WineQuality$red

}
\keyword{datasets}
